Update CI workflow to run only affected tests and builds
In this commit, some changes have been made in the continuous integration process in the .github/workflows/ci.yml file. Instead of running all tests and builds in parallel, the command `npx nx affected` is now used. This change only runs tasks for the codebase that was affected by the latest changes, which makes the process more efficient by saving substantial computational resources and time. An additional task is also added that uses nrwl/nx-set-shas@v3 to set terms.