Add read permissions to GitHub Actions CI workflow
In this change, extra permissions have been added to the CI workflow under jobs>build configuration in the GitHub Actions CI script. The aim of this alteration is to let the workflow have read permissions to both the contents and the GitHub Actions of the repository, hence allowing a smoother and more secure execution.