Update CI workflow to deploy to GitHub Pages
The previous CI workflow used to just archive the production artifacts, but now we're deploying the frontend application directly to the GitHub Pages after building it. This uses the 'ghaction-github-pages' GitHub action, which requires certain environment secrets to be set. As such, a 'GITHUB_TOKEN' environment variable has also been added to the workflow.